src/inference_server/fastapi_requirements.txt
pandas==2.1.4
numpy==1.26.4
scipy==1.13.0
mlflow==2.12.2
fredapi==0.5.2
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.9.2
scikit-learn==1.2.2
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
hmmlearn
xgboost